add firmware directory. This directory include arduino's code.
