+++ jsk_2014_picking_challenge/firmware/CMakeLists.txt
@@ -1,0 +1,12 @@
+cmake_minimum_required(VERSION 2.8.3)
+
+include_directories(${ROS_LIB_DIR})
+
+# Remove this if using an Arduino without native USB (eg, other than Leonardo)
+add_definitions(-DUSB_CON)
+
+generate_arduino_firmware(switching_gripper
+  SRCS switching_gripper.cpp ${ROS_LIB_DIR}/time.cpp
+  BOARD atmega328
+  PORT /dev/ttyUSB0
+)
