+++ jsk_2014_picking_challenge/package.xml
@@ -23,6 +23,8 @@
   <build_depend>moveit_resources</build_depend>
   <build_depend>jsk_perception</build_depend>
   <build_depend>jsk_recognition_msgs</build_depend>
+  <build_depend>rosserial_arduino</build_depend>
+  <build_depend>rosserial_client</build_depend>
 
   <run_depend>baxtereus</run_depend>
   <run_depend>geometry_msgs</run_depend>
@@ -36,6 +38,8 @@
   <run_depend>SoundRequest</run_depend>
   <run_depend>actionlib</run_depend>
   <run_depend>actionlib_msgs</run_depend>
+  <run_depend>rosserial_arduino</run_depend>
+  <run_depend>rosserial_client</run_depend>
 
   <export>
   </export>
