--- jsk_2014_picking_challenge/src/switching_gripper.cpp
@@ -1,27 +1,0 @@
-/*
- * Switch SSR on callback
- */
-
-#include <ros.h>
-#include <std_msgs/Empty.h>
-
-ros::NodeHandle  nh;
-
-void messageCb( const std_msgs::Empty& toggle_msg){
-  digitalWrite(13, HIGH-digitalRead(13));   // Switch SSR
-}
-
-ros::Subscriber<std_msgs::Empty> sub("on_off_gripper", &messageCb );
-
-void setup()
-{
-  pinMode(13, OUTPUT);
-  nh.initNode();
-  nh.subscribe(sub);
-}
-
-void loop()
-{
-  nh.spinOnce();
-  delay(1);
-}
